+++ p2p/src/net/default_backend/peer.rs
@@ -159,7 +159,7 @@ where
         Ok(())
     }
 
-    async fn maybe_send_will_disconnect(
+    async fn maybe_send_will_disconnect_for_protocol_version(
         &mut self,
         reason: Option<DisconnectionReason>,
         peer_protocol_version: ProtocolVersion,
@@ -177,6 +177,26 @@ where
                     }))
                     .await?;
             }
+        }
+
+        Ok(())
+    }
+
+    async fn maybe_send_will_disconnect(
+        &mut self,
+        reason: Option<DisconnectionReason>,
+    ) -> crate::Result<()> {
+        if let Some(common_protocol_version) = self.common_protocol_version {
+            self.maybe_send_will_disconnect_for_protocol_version(
+                reason,
+                common_protocol_version.into(),
+            )
+            .await?;
+        } else {
+            // Getting here means that the handshake hasn't been completed yet.
+            log::debug!(
+                "self.common_protocol_version is not set when attempting to send WillDisconnect"
+            );
         }
 
         Ok(())
@@ -208,7 +228,7 @@ where
             )
         })();
 
-        self.maybe_send_will_disconnect(
+        self.maybe_send_will_disconnect_for_protocol_version(
             DisconnectionReason::from_result(&result),
             peer_protocol_version,
         )
@@ -422,7 +442,7 @@ where
                         .await;
                     if let Err(send_error) = send_result {
                         log::error!(
-                            "Cannot send PeerEvent::MisbehavedOnHandshake to peer {}: {}",
+                            "Cannot send PeerEvent::MisbehavedOnHandshake for peer {}: {}",
                             self.peer_id,
                             send_error
                         );
@@ -459,14 +479,7 @@ where
                     BackendEvent::SendMessage(message) => self.socket.send(*message).await?,
                     BackendEvent::Disconnect {reason} => {
                         log::debug!("Disconnection requested for peer {}, the reason is {:?}", self.peer_id, reason);
-                        if let Some(common_protocol_version) = self.common_protocol_version {
-                            self.maybe_send_will_disconnect(reason, common_protocol_version.into()).await?;
-                        } else {
-                            // Getting here means that we've got a disconnection request when
-                            // the handshake hasn't been completed yet.
-                            log::debug!("self.common_protocol_version is not set when BackendEvent::Disconnect is received");
-                        }
-
+                        self.maybe_send_will_disconnect(reason).await?;
                         return Ok(());
                     },
                 },
@@ -482,7 +495,26 @@ where
                         ).await?;
                     }
                     Err(err) => {
-                        log::info!("Connection closed for peer {}, reason {err:?}", self.peer_id);
+                        log::info!("Connection closed for peer {}, reason: {err:?}", self.peer_id);
+
+                        let err = P2pError::NetworkingError(err);
+                        self.maybe_send_will_disconnect(DisconnectionReason::from_error(&err)).await?;
+
+                        let ban_score = err.ban_score();
+                        if ban_score > 0 {
+                            let send_result = self
+                                .peer_event_sender
+                                .send(PeerEvent::Misbehaved { error: err })
+                                .await;
+                            if let Err(send_error) = send_result {
+                                log::error!(
+                                    "Cannot send PeerEvent::Misbehaved for peer {}: {}",
+                                    self.peer_id,
+                                    send_error
+                                );
+                            }
+                        }
+
                         return Ok(());
                     }
                 }